separated gameroutes file into publisherRoutes and generalRoute
